requirements.txt
fastapi==0.115.0
psycopg2-binary==2.9.10
python-multipart==0.0.20
httpx==0.28.1
pytest-asyncio==0.24.0
pytest==8.3.4
scikit-learn==1.6.0
pandas==2.2.3
numpy==2.2.1
transformers==4.47.1
torch==2.7.1
python-dotenv==1.0.1
alembic==1.14.0
sqlalchemy==2.0.36
pydantic-settings==2.6.1
pydantic[email]==2.10.3
uvicorn[standard]==0.34.0
python-jose[cryptography]==3.3.0

